refactor: move cli to module and seperate helpers from commands
