+++ pyproject.toml
@@ -33,4 +33,4 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-brisk = "brisk.cli:cli"
+brisk = "brisk.cli.cli:cli"
